requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
requests==2.31.0
pydantic==2.4.2
python-dotenv==1.0.0
google-generativeai==0.3.1
Pillow==10.1.0
httpx==0.24.1
